backend/requirements.txt
# Web framework
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Settings & validation
pydantic-settings>=2.0.0
email-validator>=2.0.0

# Database (async MongoDB)
motor>=3.3.0

# HTTP client
httpx>=0.27.0

# AI / LLM
langchain>=0.2.0
langchain-core>=0.2.0
langchain-groq>=0.1.0
langchain-community>=0.2.0

# Vector store
chromadb>=0.5.0

# PDF processing
pdfplumber>=0.10.0
PyPDF2>=3.0.0
